ENH: refs #0336. Execute works, added logoroll and Execute string.
Changed the execution, now Seedpoints come from an itemrevision, and the
underlying batchmake infrastructure supports the execution and return.
Still need a great deal of work cleaning up the code, and on front end
for setting up the workflow and displaying the results.
CSS changes from Springer Interactive have been included in qibench.css,
which should be a much better way to work together on development.  The
string "Execute" has also been added to the application.